# public/doc/swagger.yaml
swagger: '2.0'
info:
  description: Asset DB ドキュメント
  version: 0.0.1
  title: Asset DB REST APIs
#host: localhost:8080
basePath: /api/db
tags:
  - name: assets
    description: Assets APIs
schemes:
  - http
  - https
paths:
  /create_asset:
    post:
      tags:
        - assets
      summary: Asset 作成
      description: Asset 作成
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/AssetRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /read_asset/{asset_id}:
    get:
      tags:
        - assets
      summary: Asset 取得
      description: Asset 取得
      produces:
        - application/json
      parameters:
        - name: asset_id
          type: string
          in: path
          description: Asset ID
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /read_assets:
    get:
      tags:
        - assets
      summary: Assets 全取得
      description: Assets 全取得
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /plus_asset:
    post:
      tags:
        - assets
      summary: Asset 補給
      description: Asset 補給
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/AssetRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /minus_asset:
    post:
      tags:
        - assets
      summary: Asset 減少
      description: Asset 減少
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/AssetRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /consume_asset_percent/{asset_id}:
    post:
      tags:
        - assets
      summary: Asset 利用
      description: Asset 利用
      produces:
        - application/json
      parameters:
        - name: asset_id
          type: string
          in: path
          description: Asset ID
          required: true
        - name: body
          in: body
          schema:
            $ref: '#/definitions/PercentRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /consume_asset_amount/{asset_id}:
    post:
      tags:
        - assets
      summary: Asset 利用
      description: Asset 利用
      produces:
        - application/json
      parameters:
        - name: asset_id
          type: string
          in: path
          description: Asset ID
          required: true
        - name: body
          in: body
          schema:
            $ref: '#/definitions/AmountRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /transport_asset/{asset_id_from}/{asset_id_to}:
    post:
      tags:
        - assets
      summary: Asset 移動
      description: Asset 移動
      produces:
        - application/json
      parameters:
        - name: asset_id_from
          type: string
          in: path
          description: 移動元 Asset ID
          required: true
        - name: asset_id_to
          type: string
          in: path
          description: 移動先 Asset ID
          required: true
        - name: body
          in: body
          schema:
            $ref: '#/definitions/AmountRequest'
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /reset_asset/{asset_id}:
    post:
      tags:
        - assets
      summary: Asset リセット
      description: Asset リセット
      produces:
        - application/json
      parameters:
        - name: asset_id
          type: string
          in: path
          description: Asset ID
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /reset_assets:
    post:
      tags:
        - assets
      summary: Asset 全リセット
      description: Asset 全リセット
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /remove_asset/{asset_id}:
    delete:
      tags:
        - assets
      summary: Asset 削除
      description: Asset 削除
      produces:
        - application/json
      parameters:
        - name: asset_id
          type: string
          in: path
          description: Asset ID
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /remove_assets:
    delete:
      tags:
        - assets
      summary: Asset 全削除
      description: Asset 全削除
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
definitions:
  AssetRequest:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      name_a:
        type: string
      num_a:
        type: number
      name_b:
        type: string
      num_b:
        type: number
  PercentRequest:
    type: object
    properties:
      percent:
        type: number
  AmountRequest:
    type: object
    properties:
      amount:
        type: number